Added code to make the navigation arrows in Romania.java display correctly for RTL as well as LTR product flavors. Refactored Testing.java as Settings.java.
