+++ app/src/main/java/org/alphatilesapps/alphatiles/Settings.java
@@ -1,0 +1,8 @@
+package org.alphatilesapps.alphatiles;
+
+public class Settings
+{
+	public static boolean tempSoundPoolSwitch = true;    // true for SoundPool, false for MediaPlayer
+	public static boolean forceRTL = true;		//use if aa_langinfo.txt has RTL for Script Direction
+
+}
