--- app/src/main/java/org/alphatilesapps/alphatiles/Testing.java
@@ -1,7 +1,0 @@
-package org.alphatilesapps.alphatiles;
-
-public class Testing
-{
-	public static boolean tempSoundPoolSwitch = true;	// true for SoundPool, false for MediaPlayer
-	public static boolean forceRTL = true;
-}
