+++ app/src/main/res/layout/romania.xml
@@ -27,7 +27,7 @@
         android:layout_width="0dp"
         android:layout_height="0dp"
         android:layout_marginEnd="5dp"
-        android:onClick="goBackOneTile"
+        android:onClick="goToTileOnTheLeft"
         app:layout_constraintBottom_toTopOf="@+id/guidelineHMagTileTop"
         app:layout_constraintEnd_toStartOf="@+id/tileBoxTextView"
         app:layout_constraintHorizontal_bias="1.0"
@@ -36,21 +36,23 @@
         app:layout_constraintVertical_bias="0.0"
         android:contentDescription="@string/goBackward"
         app:srcCompat="@drawable/zz_backward"
-        android:layout_marginRight="5dp" />
+        android:layout_marginRight="5dp"
+        android:rotationY="@integer/mirror_flip"/>
 
     <ImageView
         android:id="@+id/rightArrowImage"
         android:layout_width="0dp"
         android:layout_height="0dp"
         android:layout_marginStart="5dp"
-        android:onClick="goForwardOneTile"
+        android:onClick="goToTileOnTheRight"
         app:layout_constraintBottom_toTopOf="@+id/guidelineHMagTileTop"
         app:layout_constraintEnd_toStartOf="@+id/guidelineV4"
         app:layout_constraintStart_toEndOf="@+id/tileBoxTextView"
         app:layout_constraintTop_toTopOf="@+id/guidelineH1andHalf"
         app:srcCompat="@drawable/zz_forward"
         android:contentDescription="@string/goForward"
-        android:layout_marginLeft="5dp" />
+        android:layout_marginLeft="5dp"
+        android:rotationY="@integer/mirror_flip"/>
 
     <ImageView
         android:id="@+id/wordImage"
@@ -77,6 +79,7 @@
         app:layout_constraintStart_toStartOf="@+id/guidelineVMagLeft"
         app:layout_constraintTop_toTopOf="@+id/guidelineHMagTop"
         app:srcCompat="@drawable/zz_magnifying_glass"
+        android:rotationY="@integer/mirror_flip"
         android:contentDescription="@string/seeMoreOfSameGameTile" />
 
     <TextView
